Add hook to augment `Platform` objects for Artifact selection
This, combined with the recent overhaul of BinaryPlatforms to give
extended tags to `Platform` objects, enables packages to run arbitrary
(ideally, small) pieces of code to better inform artifact selection.

`Platform`-specific Artifacts can now be tagged with arbitrary metadata,
examples of which are `libc` version, processor microarchitecture, or
CUDA version.  Because we do not want to have to teach Julia how to
autodetect its compatibility with all possible tags, we instead add the
ability for a package to perform the autodetection, and run that code
after the package is installed, but before its artifacts are downloaded.

Cooperating packages will create a `.pkg/select_artifacts_hook.jl`
file within their package, and running that file will print a TOML-
encoded dictionary containing mappings for each artifact that should be
installed.  Output should mimic what is returned by the `Artifacts`
function `select_downloadable_artifacts()`, and in fact, it is encouraged
to organize package code into two pieces; a platform augmentation function
that is then used to pass an augmented platform to
`select_downloadable_artifacts()`, whose output is then TOML-encoded and
printed to standard output.  This is demonstrated in the
`AugmentedPlatformHook` test package within `test/test_packages`.

Note that when those packages use the `@artifact_str()` macro themselves,
they should pass in an already-augmented `Platform` object as the second
argument to `@artifact_str`.  To make this easier, we suggest mimicking
the organization of the `AugmentedPlatformHook` package, where the
platform augmentation function can be imported by both the
`select_artifacts_hook.jl` script and the main package.